Interestingly averaging sale prices fixes FarmersAndWorker scenario as the firm gets a better idea of what the real price is (that is, the price is really 2.7 and the PID by pulsating between 2 and 3 signals it). Unfortunately averaging creates all sorts of issues in the competitive market. Creates both higher prices and higher standard deviation. I have to figure this out before I can push.

+++ src/main/java/agents/firm/purchases/prediction/AroundShockLinearRegressionPurchasePredictor.java
@@ -79,7 +79,7 @@ public class AroundShockLinearRegressionPurchasePredictor implements PurchasesPr
      * @return the predicted price or -1 if there are no predictions.
      */
     @Override
-    public int predictPurchasePriceWhenNoChangeInProduction(PurchasesDepartment dept) {
+    public float predictPurchasePriceWhenNoChangeInProduction(PurchasesDepartment dept) {
         return predictor.predictPurchasePriceWhenNoChangeInProduction(dept);
     }
 
@@ -101,7 +101,7 @@ public class AroundShockLinearRegressionPurchasePredictor implements PurchasesPr
      * @return the predicted price or -1 if there are no predictions.
      */
     @Override
-    public int predictPurchasePriceWhenIncreasingProduction(PurchasesDepartment dept) {
+    public float predictPurchasePriceWhenIncreasingProduction(PurchasesDepartment dept) {
 
         updateModelIfPossible(dept);
 
@@ -199,7 +199,7 @@ public class AroundShockLinearRegressionPurchasePredictor implements PurchasesPr
      * @return the predicted price or -1 if there are no predictions.
      */
     @Override
-    public int predictPurchasePriceWhenDecreasingProduction(PurchasesDepartment dept) {
+    public float predictPurchasePriceWhenDecreasingProduction(PurchasesDepartment dept) {
         updateModelIfPossible(dept);
     //    System.out.println("slope: " + predictor.getIncrementDelta());
