+++ src/main/java/agents/firm/sales/prediction/FixedPriceFakePredictor.java
@@ -46,7 +46,7 @@ public class FixedPriceFakePredictor extends BaseSalesPredictor implements Purch
      * @return the best offer available/predicted or -1 if there are no quotes/good predictions
      */
     @Override
-    public int predictSalePriceAfterIncreasingProduction(SalesDepartment dept, int expectedProductionCost, int increaseStep) {
+    public float predictSalePriceAfterIncreasingProduction(SalesDepartment dept, int expectedProductionCost, int increaseStep) {
         return priceToPredict;
     }
 
@@ -59,7 +59,7 @@ public class FixedPriceFakePredictor extends BaseSalesPredictor implements Purch
      * @return the best offer available/predicted or -1 if there are no quotes/good predictions
      */
     @Override
-    public int predictSalePriceAfterDecreasingProduction(SalesDepartment dept, int expectedProductionCost, int decreaseStep) {
+    public float predictSalePriceAfterDecreasingProduction(SalesDepartment dept, int expectedProductionCost, int decreaseStep) {
         return priceToPredict;
     }
 
@@ -71,7 +71,7 @@ public class FixedPriceFakePredictor extends BaseSalesPredictor implements Purch
      * @return predicted price
      */
     @Override
-    public int predictSalePriceWhenNotChangingProduction(SalesDepartment dept) {
+    public float predictSalePriceWhenNotChangingProduction(SalesDepartment dept) {
         return priceToPredict;
     }
 
@@ -83,7 +83,7 @@ public class FixedPriceFakePredictor extends BaseSalesPredictor implements Purch
      * @return the predicted price or -1 if there are no predictions.
      */
     @Override
-    public int predictPurchasePriceWhenIncreasingProduction(PurchasesDepartment dept) {
+    public float predictPurchasePriceWhenIncreasingProduction(PurchasesDepartment dept) {
         return priceToPredict;
     }
 
@@ -94,7 +94,7 @@ public class FixedPriceFakePredictor extends BaseSalesPredictor implements Purch
      * @return the predicted price or -1 if there are no predictions.
      */
     @Override
-    public int predictPurchasePriceWhenDecreasingProduction(PurchasesDepartment dept) {
+    public float predictPurchasePriceWhenDecreasingProduction(PurchasesDepartment dept) {
         return priceToPredict;
     }
 
@@ -105,7 +105,7 @@ public class FixedPriceFakePredictor extends BaseSalesPredictor implements Purch
      * @return the predicted price or -1 if there are no predictions.
      */
     @Override
-    public int predictPurchasePriceWhenNoChangeInProduction(PurchasesDepartment dept) {
+    public float predictPurchasePriceWhenNoChangeInProduction(PurchasesDepartment dept) {
         return priceToPredict;
     }
