+++ src/main/java/agents/firm/sales/prediction/LearningDecreaseWithTimeSeriesSalesPredictor.java
@@ -138,7 +138,7 @@ public class LearningDecreaseWithTimeSeriesSalesPredictor
      * @return the best offer available/predicted or -1 if there are no quotes/good predictions
      */
     @Override
-    public int predictSalePriceAfterIncreasingProduction(SalesDepartment dept, int expectedProductionCost, int increaseStep) {
+    public float predictSalePriceAfterIncreasingProduction(SalesDepartment dept, int expectedProductionCost, int increaseStep) {
         Preconditions.checkArgument(increaseStep >= 0);
 
         if(dept.numberOfObservations() > 200 &&
@@ -174,7 +174,7 @@ public class LearningDecreaseWithTimeSeriesSalesPredictor
      * @return the best offer available/predicted or -1 if there are no quotes/good predictions
      */
     @Override
-    public int predictSalePriceAfterDecreasingProduction(SalesDepartment dept, int expectedProductionCost, int decreaseStep)
+    public float predictSalePriceAfterDecreasingProduction(SalesDepartment dept, int expectedProductionCost, int decreaseStep)
     {
         Preconditions.checkArgument(decreaseStep >= 0);
 
@@ -509,7 +509,7 @@ public class LearningDecreaseWithTimeSeriesSalesPredictor
      * @return predicted price
      */
     @Override
-    public int predictSalePriceWhenNotChangingProduction(SalesDepartment dept) {
+    public float predictSalePriceWhenNotChangingProduction(SalesDepartment dept) {
         return predictor.predictSalePriceWhenNotChangingProduction(dept);
     }
