+++ src/main/java/agents/firm/sales/prediction/LookupSalesPredictor.java
@@ -37,7 +37,7 @@ public class LookupSalesPredictor extends BaseSalesPredictor {
      * @return the best offer available/predicted or -1 if there are no quotes/good predictions
      */
     @Override
-    public int predictSalePriceAfterIncreasingProduction(SalesDepartment dept, int expectedProductionCost, int increaseStep) {
+    public float predictSalePriceAfterIncreasingProduction(SalesDepartment dept, int expectedProductionCost, int increaseStep) {
 
         return (int) Math.round(dept.getMarket().getLatestObservation(MarketDataType.AVERAGE_CLOSING_PRICE));
 
@@ -52,7 +52,7 @@ public class LookupSalesPredictor extends BaseSalesPredictor {
      * @return
      */
     @Override
-    public int predictSalePriceAfterDecreasingProduction(SalesDepartment dept, int expectedProductionCost, int decreaseStep) {
+    public float predictSalePriceAfterDecreasingProduction(SalesDepartment dept, int expectedProductionCost, int decreaseStep) {
 
         return Math.round(dept.getMarket().getLastDaysAveragePrice());
     }
@@ -82,7 +82,7 @@ public class LookupSalesPredictor extends BaseSalesPredictor {
      * @return predicted price
      */
     @Override
-    public int predictSalePriceWhenNotChangingProduction(SalesDepartment dept) {
+    public float predictSalePriceWhenNotChangingProduction(SalesDepartment dept) {
         return Math.round(dept.getMarket().getLastDaysAveragePrice());
 
     }
