+++ src/main/java/agents/firm/sales/prediction/RegressionSalePredictor.java
@@ -79,7 +79,7 @@ public class RegressionSalePredictor extends BaseSalesPredictor{
      * @return the best offer available/predicted or -1 if there are no quotes/good predictions
      */
     @Override
-    public int predictSalePriceAfterIncreasingProduction(SalesDepartment dept, int expectedProductionCost, int increaseStep)
+    public float predictSalePriceAfterIncreasingProduction(SalesDepartment dept, int expectedProductionCost, int increaseStep)
     {
         Preconditions.checkArgument(increaseStep >= 0);
 
@@ -118,7 +118,7 @@ public class RegressionSalePredictor extends BaseSalesPredictor{
      * @return the best offer available/predicted or -1 if there are no quotes/good predictions
      */
     @Override
-    public int predictSalePriceAfterDecreasingProduction(SalesDepartment dept, int expectedProductionCost, int decreaseStep) {
+    public float predictSalePriceAfterDecreasingProduction(SalesDepartment dept, int expectedProductionCost, int decreaseStep) {
         Preconditions.checkArgument(decreaseStep >= 0);
         //regress and return
         updateModel();
@@ -313,7 +313,7 @@ public class RegressionSalePredictor extends BaseSalesPredictor{
      * @return predicted price
      */
     @Override
-    public int predictSalePriceWhenNotChangingProduction(SalesDepartment dept) {
+    public float predictSalePriceWhenNotChangingProduction(SalesDepartment dept) {
         //regress and return
         updateModel();
