+++ src/main/java/agents/firm/sales/prediction/SamplingLearningDecreaseSalesPredictor.java
@@ -83,7 +83,7 @@ public class SamplingLearningDecreaseSalesPredictor extends AbstractWorkerLearni
      * @return the best offer available/predicted or -1 if there are no quotes/good predictions
      */
     @Override
-    public int predictSalePriceAfterIncreasingProduction(SalesDepartment dept, int expectedProductionCost, int increaseStep) {
+    public float predictSalePriceAfterIncreasingProduction(SalesDepartment dept, int expectedProductionCost, int increaseStep) {
         this.department=dept;
 
 
@@ -104,7 +104,7 @@ public class SamplingLearningDecreaseSalesPredictor extends AbstractWorkerLearni
      * @return the best offer available/predicted or -1 if there are no quotes/good predictions
      */
     @Override
-    public int predictSalePriceAfterDecreasingProduction(SalesDepartment dept, int expectedProductionCost, int decreaseStep) {
+    public float predictSalePriceAfterDecreasingProduction(SalesDepartment dept, int expectedProductionCost, int decreaseStep) {
         this.department=dept;
 
 
@@ -124,7 +124,7 @@ public class SamplingLearningDecreaseSalesPredictor extends AbstractWorkerLearni
      * @return predicted price
      */
     @Override
-    public int predictSalePriceWhenNotChangingProduction(SalesDepartment dept) {
+    public float predictSalePriceWhenNotChangingProduction(SalesDepartment dept) {
         return predictor.predictSalePriceWhenNotChangingProduction(dept);
 
     }
